Redirect customer from checkout  to order pay
The current implementation redirects the customer to the plugin customs template, and ends the process there, which results in other templates not being rendered unless they were added earlier in the template pipeline.

With this change, we redirect the customer to the order pay page without terminating the template pipeline.